ground markers: Allow different colored markers
This adds a color indicator per marker, saved to config, allowing
multiple markers to be different colors. In addition, a configuration
option is added to switch between per-tile color display or current
configured color display. (that is, new or existing behavior)

This change will set a color in config for old markers with null colors
when those markers are loaded.

Fixes runelite/runelite#3395