config: add tag suggestions
This is intended to make the tags feature more discoverable, as people
keep requesting "plugin categories". Putting plugins into a singular
category that will make sense for most users is impossible, as many
plugins will can logically be placed into multiple categories.

The Category Tag list is intended to be a short, curated, list of tags
that are helpful to a user trying to find plugins.