Apply default font to text with unicode characters
MacOS does not support fallback fonts, and any character not in our RS
fonts do not render correctly.

We only render unicode characters a handful of places, mostly for the
check mark/cross in overlays, and on the icon text field suggestion
button. So this sets the font of those places to the default system font
which can render them correctly.