discord: fix npe from menu event timing out with elapsed total time
Setting the start time to null isn't necessary, since IN_MENU is marked
shouldRestart, it always resets the start time when the status is changed to
menu. This was clearing it, forcing a presence update, and then erroring when
trying to compute the original start time. It also would have never really unset,
even if it wasn't erroring later, due to IN_MENU also being unclearable, and
so would have never been removed in the first place.

Additionally the status reset test is also wrong since it is testing the end
timestamp which has never been used by us, making the test useless.