barrows: fix vanilla overlay flashing with unlocked fps
The flashing is caused by the vanilla overlays being hidden by the
detached overlay after being drawn, and so the remaining frames before
the next tick would have them hidden. Just always hide the overlay in
BeforeRender so that they are never drawn when the plugin is on.

Also add a quick check if the barrows region is loaded before trying to
draw the barrows brothers minimap overlay.