Set macos quitStrategy to CLOSE_ALL_WINDOWS
The quit strategy API existed in com.apple.eawt prior to Java 9, and was moved
into java.desktop in 9. This move makes our orange-extensions API incompatiable
with the 9 eawt package too, so we can not use Application.setQuitStrategy on
9+ either, without making orange-extensions multi-release too.

Since Desktop.setQuitStrategy is the new standard way of doing this, use it if
available, and otherwise use the Java 8 Application.setQuitStrategy instead.