clues: Fix named object clue plane change scan
The previous code to scan for named objects neglected to clear the set
of named objects to be marked, leading to multilocs sometimes being
incorrectly marked after plane changes. This commit fixes that, and
cleans up the named object scan code within onGameTick as there should
never be a need to both change the plane and mark the next tick to scan
for named objects and to perform said scan within the same tick.