spec counter: fix tracking dorgeshuun specs
Correct both the cycle delay computation and the server tick conversion
computation.

Note that the server tick counter in both hitsplat applied and varbit
changed is prior to the tick counter being updated, which happens on
the following cycle to npc update, so it is safe to compare the two
tick values there. Using the tick value in the invokeLater callback
is too late as it is after it has been incremented.

The hitsplat delay computation needs to include the initial projectile
cycle delay as the var changes at the same time the projectile spawns,
and not first moves.

The server does not round the delay calculation and instead uses it as
an offset from the next tick.