gpu: fix amd miscompile of frag.glsl
amd drivers 22.7.1 and later miscompile a switch with two leading cases when one falls through.

The switch was originally put in place due to intel drivers breaking with a chain of ifs, which was fixed in ad4d1787a15a733ee2883c140055b0d53ba3f6a7.

Hopefully this works better since there is an else {} case and also the color blind code is not behind a condition anymore.