raids: reset raid on party id change
This doesn't happen normally. There is a race which allows creating
multiple parties in the same fc in the same world, by resetting the raid
and then on a 2nd account interacting with the board to quickly make a
new raid at the same time.

In this state, when the raid is reset, it resets the party id to the
other party for the fc like this:

partyid = other
in raid = false
in raid = true
partyid = -1
partyid = new

Instead of the usual reset which looks like this:

partyid = -1
in raid = false
in raid = true
partyid = new

So detect changing between party ids and reset the raid then.

Changing from party id -> -1 is valid and happens when a raid starts,
and changing from -1 -> party id happens after in_raid is flipped and
the raid has been scouted.