gpu: fix tile uvs
The client passes different local coordinates to the texture rasterizer
depending on if the tile is flat (all corners have the height) or not.

The code we've had this entire time was always assuming flat tiles. This
used to kind of work, but with the recent changes to compute the uvs in
the shader to correct for textris, it has made non-flat tiles more
broken.