clues: recognize (l)(t) variant of dragon defender
Right now, the clue plugin does not recognize the locked (l) and trimmed
(t) variant of the Dragon Defender as fulfilling the requirement for
Falo the Bard's Dragon or Avernic Defender clue.

This is incorrect, as the Dragon Defender (l)(t) is accepted by Falo.
Let's update the requirement collection such that this is no longer the
case.

This fixes issue #17422.

Signed-off-by: Geordan Neukum <gneukum1@gmail.com>