gpu: fix priority face average calculation
While the client defaults the averages to 0, it also offsets the
distances by the model radius, meaning 0 is always the closest distance.

This is a regression from 56b8954861f7046bbc31f872623b5a934577bd6b which
removed the radius offset from the distances.

Initialize averages to -6000 because the model max diameter is 6000 and
so it should always be less than the computed distance.