loottracker: add araxxor
THis has a unique drop mechanic of course, where after harvesting the
npc you receive the loot. There are no additional chat messages. I am
using the harvest animation instead to detect it.

Additionally there is a "Destroy" option to destroy loot, but still do
the harvest animation, so I've added a timeout to the inv change
detector to detect this.